+++ src/Entities/Reporting/TransactionSummary.php
@@ -4,6 +4,7 @@ namespace GlobalPayments\Api\Entities\Reporting;
 use GlobalPayments\Api\Entities\AlternativePaymentResponse;
 use GlobalPayments\Api\Entities\BankPaymentResponse;
 use GlobalPayments\Api\Entities\Enums\PaymentEntryMode;
+use GlobalPayments\Api\Entities\LodgingData;
 
 class TransactionSummary
 {